requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1  
Django==5.1.6
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
PyMySQL==1.1.1
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3
python-dotenv==1.0.1
django-autoslug==1.9.8


